+++ lib/glimmer/libui/quit_menu_item_proxy.rb
@@ -52,7 +52,7 @@ module Glimmer
       def build_control
         @libui = @parent_proxy.append_quit_item(*@args)
         handle_listener('on_clicked') do
-          destroy # TODO look into finding a way to destroy main window
+          ControlProxy.main_window_proxy&.destroy
           ::LibUI.quit
           0
         end