+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 
 - Make `padded 1` the default in `horizontal_box` and `vertical_box` to achieve nicer looking GUI by default
 - Make `margined 1` the default in `group` to achieve nicer looking GUI by default
+- Destroy main window upon hitting quit on quit menu item
+- Rename `ControlProxy::all_controls` to `ControlProxy::all_control_proxies` to more accurately describe its contents
+- Add `ControlProxy::main_window_proxy` method to retrieve main window proxy
 
 ## 0.0.6